requirements.txt
streamlit
pdfplumber
rank_bm25
groq
python-dotenv
